requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
aiosqlite
aiosmtplib
pydantic-settings
tzlocal
cryptography
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
